+++ plugins/catalog-backend-module-msgraph/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-catalog-backend-module-msgraph
 
+## 0.6.6-next.1
+
+### Patch Changes
+
+- 29a4aa8: fix(config): add missing parameters in config schema
+
+## 0.6.6-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.1.1-next.0
+  - @backstage/catalog-model@1.7.2
+  - @backstage/config@1.3.1
+  - @backstage/plugin-catalog-common@1.1.2
+  - @backstage/plugin-catalog-node@1.15.1-next.0
+
 ## 0.6.5
 
 ### Patch Changes