+++ plugins/notifications-node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-notifications-node
 
+## 0.2.11-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.1.1-next.0
+  - @backstage/catalog-client@1.9.0
+  - @backstage/catalog-model@1.7.2
+  - @backstage/plugin-notifications-common@0.0.7
+  - @backstage/plugin-signals-node@0.1.16-next.0
+
 ## 0.2.10
 
 ### Patch Changes